2023/projects/01_UniLock/code/unilock_app_docker/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
idna==3.4
pypng==0.20220715.0
qrcode==7.4.2
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
python-decouple==3.5
python-dotenv==1.0.0
paho-mqtt==1.6.1
